# oss-migration-guide/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coffee Orders API",
    "description": "A CRUD API for managing coffee orders and available coffee types.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Development server" }
  ],
  "paths": {
    "/orders": {
      "get": {
        "summary": "Get Orders",
        "description": "Retrieve all coffee orders.\nIf 'coffee_type' is provided, returns orders matching that coffee type.",
        "operationId": "get_orders_orders_get",
        "parameters": [
          {
            "name": "coffee_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Optional filter by coffee type (case-insensitive)",
              "title": "Coffee Type"
            },
            "description": "Optional filter by coffee type (case-insensitive)",
            "examples": {
              "latte": {
                "summary": "Filter Latte orders",
                "value": "Latte"
              },
              "espresso": {
                "summary": "Filter Espresso orders",
                "value": "Espresso"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CoffeeOrder" },
                  "title": "Response Get Orders Orders Get"
                },
                "examples": {
                  "multiple_orders": {
                    "summary": "List of coffee orders",
                    "value": [
                      {
                        "id": 1,
                        "customer_name": "Alice",
                        "coffee_type": "Latte",
                        "size": "Medium",
                        "extras": ["Extra shot", "Soy milk"],
                        "price": 4.50
                      },
                      {
                        "id": 2,
                        "customer_name": "Bob",
                        "coffee_type": "Espresso",
                        "size": "Small",
                        "extras": ["Extra shot"],
                        "price": 3.50
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Order",
        "description": "Create a new coffee order.\nValidates that the coffee type exists.",
        "operationId": "create_order_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CoffeeOrder" },
              "examples": {
                "simple_order": {
                  "summary": "Basic coffee order",
                  "value": {
                    "id": 3,
                    "customer_name": "Charlie",
                    "coffee_type": "Americano",
                    "size": "Large",
                    "price": 3.75
                  }
                },
                "complex_order": {
                  "summary": "Coffee order with extras",
                  "value": {
                    "id": 4,
                    "customer_name": "Diana",
                    "coffee_type": "Cappuccino",
                    "size": "Medium",
                    "extras": ["Whipped cream", "Caramel syrup"],
                    "price": 5.25
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeOrder" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}": {
      "get": {
        "summary": "Get Order",
        "description": "Retrieve a specific coffee order by its ID.",
        "operationId": "get_order_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Order Id" },
            "examples": {
              "order1": {
                "summary": "First order",
                "value": 1
              },
              "order2": {
                "summary": "Second order",
                "value": 2
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeOrder" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Order",
        "description": "Update an existing coffee order.",
        "operationId": "update_order_orders__order_id__put",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Order Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CoffeeOrderUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeOrder" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Order",
        "description": "Delete a coffee order.",
        "operationId": "delete_order_orders__order_id__delete",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Order Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/coffee-types": {
      "get": {
        "summary": "Get Coffee Types",
        "description": "Retrieve all available coffee types.",
        "operationId": "get_coffee_types_coffee_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/CoffeeType" },
                  "type": "array",
                  "title": "Response Get Coffee Types Coffee Types Get"
                },
                "examples": {
                  "coffee_types": {
                    "summary": "List of available coffee types",
                    "value": [
                      {
                        "id": 1,
                        "name": "Espresso",
                        "description": "Strong and bold coffee shot",
                        "price_multiplier": 1.0
                      },
                      {
                        "id": 2,
                        "name": "Latte",
                        "description": "Espresso with steamed milk",
                        "price_multiplier": 1.2
                      },
                      {
                        "id": 3,
                        "name": "Mocha",
                        "description": "Espresso with chocolate and steamed milk",
                        "price_multiplier": 1.3
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Coffee Type",
        "description": "Create a new coffee type.",
        "operationId": "create_coffee_type_coffee_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CoffeeType" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeType" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/coffee-types/{type_id}": {
      "get": {
        "summary": "Get Coffee Type",
        "description": "Retrieve a specific coffee type by its ID.",
        "operationId": "get_coffee_type_coffee_types__type_id__get",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Type Id" },
            "examples": {
              "espresso": {
                "summary": "Espresso type ID",
                "value": 1
              },
              "latte": {
                "summary": "Latte type ID",
                "value": 2
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeType" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Coffee Type",
        "description": "Update an existing coffee type.",
        "operationId": "update_coffee_type_coffee_types__type_id__put",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Type Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CoffeeType" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CoffeeType" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Coffee Type",
        "description": "Delete a coffee type.",
        "operationId": "delete_coffee_type_coffee_types__type_id__delete",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Type Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoffeeOrder": {
        "properties": {
          "id": { "type": "integer", "title": "Id", "example": 1 },
          "customer_name": {
            "type": "string",
            "title": "Customer Name",
            "example": "Alice"
          },
          "coffee_type": {
            "type": "string",
            "title": "Coffee Type",
            "example": "Latte"
          },
          "size": { "type": "string", "title": "Size", "example": "Medium" },
          "extras": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Extras",
            "example": ["Extra shot", "Soy milk"]
          },
          "price": { "type": "number", "title": "Price", "example": 4.5 }
        },
        "type": "object",
        "required": ["id", "customer_name", "coffee_type", "size", "price"],
        "title": "CoffeeOrder",
        "examples": [
          {
            "id": 5,
            "customer_name": "Eve",
            "coffee_type": "Mocha",
            "size": "Large",
            "extras": ["Whipped cream", "Chocolate syrup"],
            "price": 6.00
          },
          {
            "id": 6,
            "customer_name": "Grace",
            "coffee_type": "Cold Brew",
            "size": "Medium",
            "extras": ["Vanilla syrup"],
            "price": 5.50
          }
        ]
      },
      "CoffeeOrderUpdate": {
        "properties": {
          "customer_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Name",
            "example": "Alice"
          },
          "coffee_type": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Coffee Type",
            "example": "Cappuccino"
          },
          "size": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Size",
            "example": "Large"
          },
          "extras": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Extras",
            "example": ["Whipped cream"]
          },
          "price": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Price",
            "example": 5.0
          }
        },
        "type": "object",
        "title": "CoffeeOrderUpdate",
        "examples": [
          {
            "customer_name": "Frank",
            "size": "Small",
            "extras": ["Sugar-free syrup"]
          },
          {
            "coffee_type": "Cappuccino",
            "extras": ["Cinnamon", "Whipped cream"],
            "price": 5.75
          }
        ]
      },
      "CoffeeType": {
        "properties": {
          "id": { "type": "integer", "title": "Id", "example": 1 },
          "name": { "type": "string", "title": "Name", "example": "Latte" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description",
            "example": "A milk-based espresso coffee"
          },
          "price_multiplier": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Price Multiplier",
            "default": 1.0,
            "example": 1.2
          }
        },
        "type": "object",
        "required": ["id", "name"],
        "title": "CoffeeType",
        "examples": [
          {
            "id": 4,
            "name": "Cold Brew",
            "description": "Smooth, cold-steeped coffee",
            "price_multiplier": 1.4
          },
          {
            "id": 5,
            "name": "Flat White",
            "description": "Espresso with steamed whole milk",
            "price_multiplier": 1.25
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "examples": [
          {
            "detail": [
              {
                "loc": ["body", "coffee_type"],
                "msg": "Invalid coffee type specified",
                "type": "value_error"
              }
            ]
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "examples": [
          {
            "loc": ["body", "price"],
            "msg": "Price must be greater than 0",
            "type": "value_error.number.not_gt"
          }
        ]
      }
    }
  }
}