# oss-migration-guide/oss-sdk-typescript/model/validationError.ts
/**
 * Coffee Orders API
 * A CRUD API for managing coffee orders and available coffee types.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ValidationErrorLocInner } from './validationErrorLocInner';

export class ValidationError {
    'loc': Array<ValidationErrorLocInner>;
    'msg': string;
    'type': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "loc",
            "baseName": "loc",
            "type": "Array<ValidationErrorLocInner>"
        },
        {
            "name": "msg",
            "baseName": "msg",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ValidationError.attributeTypeMap;
    }
}